requirements.txt
setuptools>=65.0.0
wheel
gunicorn==21.2.0
Flask==3.0.3
numpy>=1.20,<1.25
Pillow==10.4.0
tflite-runtime==2.14.0; platform_system != "Windows"
tensorflow-cpu==2.12.0; platform_system == "Windows"
gdown